Fix bug: NECBS with bypassing conflicts
In copyConflict function, we do not copy the conflict from the parent CT node if it is an internal conflict. In findConflict function, if the CT node is newly-generated, then we copy the conflicts from its parent. If the CT node is an existing CT node with merging agents, then we do not copy conflicts from its parent CT node.